requirements.txt
pywebview==4.4.1
pyinstaller>=5.13.0
